Performs a oneway analysis

This function performs a oneway analysis of variance and returns a object of class "oneway"

oneway(formula, data)

Arguments

formula
argument that takes an dependent variable ~ independent variable

data
the data set that the dependent and independent variable will be taken from

Value

a object of class "oneway"

Examples

oneway(mpg ~ wt, mtcars)

#> 
#> Summary Statistics
#> ====================================================
#>       wt      mpg.n   mpg.mean     mpg.sd
#> 1  1.513  1.0000000 30.4000000         NA
#> 2  1.615  1.0000000 30.4000000         NA
#> 3  1.835  1.0000000 33.9000000         NA
#> 4  1.935  1.0000000 27.3000000         NA
#> 5  2.140  1.0000000 26.0000000         NA
#> 6  2.200  1.0000000 32.4000000         NA
#> 7  2.320  1.0000000 22.8000000         NA
#> 8  2.465  1.0000000 21.5000000         NA
#> 9  2.620  1.0000000 21.0000000         NA
#> 10 2.770  1.0000000 19.7000000         NA
#> 11 2.780  1.0000000 21.4000000         NA
#> 12 2.875  1.0000000 21.0000000         NA
#> 13 3.150  1.0000000 22.8000000         NA
#> 14 3.170  1.0000000 15.8000000         NA
#> 15 3.190  1.0000000 24.4000000         NA
#> 16 3.215  1.0000000 21.4000000         NA
#> 17 3.435  1.0000000 15.2000000         NA
#> 18 3.440  3.0000000 18.5666667  0.7094599
#> 19 3.460  1.0000000 18.1000000         NA
#> 20 3.520  1.0000000 15.5000000         NA
#> 21 3.570  2.0000000 14.6500000  0.4949747
#> 22 3.730  1.0000000 17.3000000         NA
#> 23 3.780  1.0000000 15.2000000         NA
#> 24 3.840  1.0000000 13.3000000         NA
#> 25 3.845  1.0000000 19.2000000         NA
#> 26 4.070  1.0000000 16.4000000         NA
#> 27 5.250  1.0000000 10.4000000         NA
#> 28 5.345  1.0000000 14.7000000         NA
#> 29 5.424  1.0000000 10.4000000         NA
#> 
#> Anova
#> ====================================================
#> 
#> Call:
#> lm(formula = formula, data = data_complete)
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -4.5432 -2.3647 -0.1252  1.4096  6.8727 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  37.2851     1.8776  19.858  < 2e-16 ***
#> wt           -5.3445     0.5591  -9.559 1.29e-10 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 3.046 on 30 degrees of freedom
#> Multiple R-squared:  0.7528,	Adjusted R-squared:  0.7446 
#> F-statistic: 91.38 on 1 and 30 DF,  p-value: 1.294e-10
#>